pyproject.toml
[tool.poetry]
name = "mdtaim"
version = "0.2.0"
description = "Multi-Dimensional TimeSeries Anomaly Detection with Itemset Mining"
authors = ["Arman Behkish"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
matrixprofile = "^1.1.10"
numpy = "1.*"
pandas = "^2.2.3"
scikit-learn = "^1.5.2"
matplotlib = "^3.9.2"
tqdm = "^4.66.5"
plotly = "^5.24.1"
pyyaml = "^6.0.2"
ipykernel = "^6.29.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
